Move Apalache client into its own module
This is almost entirely a mechanical change. We just move the Apalache
server manager and client into its own module, and rename a few types to
reflect the generality of the functionality.

This leaves `quintVerifier.ts` super slim, but all of the verify logic
(excluding the IO operations) from the `cliCommand` module should be
moved there eventually.

Closes #1396
